Split BuiltInCommand into multiple files
